+++ docs/source/changelog.rst
@@ -20,6 +20,8 @@ Added
   (`#191`_, `@bersace`_)
 * Support for eagerly raising actor exceptions in the joining thread
   with the |StubBroker|.  (`#195`_, `#203`_)
+* Support for accessing the current message from an actor via
+  |CurrentMessage|.  (`#208`_)
 
 .. _#179: https://github.com/Bogdanp/dramatiq/issues/179
 .. _#183: https://github.com/Bogdanp/dramatiq/issues/183
@@ -27,6 +29,7 @@ Added
 .. _#191: https://github.com/Bogdanp/dramatiq/pull/191
 .. _#195: https://github.com/Bogdanp/dramatiq/issues/195
 .. _#203: https://github.com/Bogdanp/dramatiq/pull/203
+.. _#208: https://github.com/Bogdanp/dramatiq/issues/208
 .. _@bersace: https://github.com/bersace
 .. _@davidt99: https://github.com/davidt99
 .. _@xelhark: https://github.com/xelhark